chore: get docs working
