+++ cspell.json
@@ -19,6 +19,8 @@
   ],
   "useGitignore": true,
   "words": [
+    "appwrite",
+    "Appwrite",
     "ARGB",
     "contador",
     "Contador",
@@ -26,12 +28,11 @@
     "Instructure",
     "localizable",
     "mostrado",
+    "nodoc",
     "página",
     "Pattonville",
     "riverpod",
-    "Texto",
-    "appwrite",
-    "Appwrite"
+    "Texto"
   ],
   "ignorePaths": [
     "lib/l10n/arb"
