+++ dartdoc_options.yaml
@@ -1,4 +1,20 @@
 dartdoc:
+  categories:
+    "Getting Started":
+      markdown: doc/getting-started.md
+      name: Getting Started
+    "C4Diagram":
+      markdown: doc/c4-diagram.md
+      name: C4 Diagram
+    "Testing":
+      markdown: doc/tests.md
+      name: Testing, Testing, 1, 2, 3
+    Translations:
+      markdown: doc/translations.md
+      name: Translations
+  categoryOrder: ["Getting Started", "Testing", "Translations", "C4Diagram"]
+  showUndocumentedCategories: true
+  includeExternal: ["test/helpers/helpers.dart"]
   linkToSource:
     uriTemplate: "https://github.com/psdtools/app/blob/main/%f%#L%l%"
     root: "."
