+++ lib/features/wrapper/presentation/wrapper_page.dart
@@ -1,3 +1,6 @@
+/// This file contains the [WrapperPage] widget, which wraps the pages.
+library pirate_code.features.wrapper.presentation.wrapper_page;
+
 import "package:auto_route/auto_route.dart";
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
