+++ lib/main.dart
@@ -1,4 +1,6 @@
 /// This is the app.
+/// {@category Getting Started}
+/// {@category C4Diagram}
 library;
 
 export "main_production.dart";
