+++ pubspec.yaml
@@ -56,6 +56,7 @@ dev_dependencies:
   build_runner: ^2.4.6
   custom_lint: ^0.4.0
   dart_define: ^2.1.0
+  dhttpd: ^4.0.1
   flutter_gen_runner: ^5.3.1
   flutter_launcher_icons: ^0.13.1
   flutter_native_splash: ^2.3.1
@@ -65,6 +66,7 @@ dev_dependencies:
   # integration_test:
   #   sdk: flutter
   freezed_lint: ^0.0.1-dev
+  mocktail: ^0.3.0
   riverpod_generator: ^2.2.4
   riverpod_lint: ^1.4.0
   very_good_analysis: ^5.0.0+1
